Revert "Merge branch 'develop' of github.com:nocox/asakatu-kun into refactor/front-store-api"
This reverts commit f12d0684acc65270400da9d69a04da331c1b15dc, reversing
changes made to fd060c60b868337fabed657d261ccc0bd2076e18.
