+++ vue/src/App.vue
@@ -21,7 +21,7 @@
 
 
 <script>
-    import Header from '@/components/Header.vue';
+    import Header from '@/components/common/Header.vue';
 
     export default {
         name: 'app',
