+++ vue/src/api/event.js
@@ -1,0 +1,1 @@
+// todo: yukikoma axios全部ここでやりたい
