+++ vue/src/components/eventList/JoinedEventList.vue
