+++ vue/src/store/modules/event/actions.js
@@ -1,0 +1,1 @@
+// todo:一旦後回し。 yukikoma
