+++ vue/src/store/modules/user/actions.js
@@ -1,0 +1,1 @@
+// todo:一旦後回し。 yukikoma
