+++ vue/src/store/modules/user/mutations.js
@@ -1,0 +1,1 @@
+// todo:一旦後回し。 yukikoma
