+++ vue/src/views/Home.vue
@@ -6,7 +6,7 @@
 
 <script>
 // @ is an alias to /src
-import HelloWorld from '@/components/HelloWorld.vue';
+import HelloWorld from '@/components/common/HelloWorld.vue';
 
 export default {
   name: 'home',
